backend/requirements.txt
# Core dependencies
setuptools>=69.0.2
wheel>=0.41.2
flask==2.3.2
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==20.1.0

# Primary dependencies
typing-extensions>=4.7.0
pydantic>=1.10.13,<2.0.0
numpy>=1.26.0

# LLM dependencies
anthropic>=0.17.0
tiktoken==0.8.0

# Vector store - using newer version
chromadb>=0.4.0

# LangChain ecosystem
langchain-core>=0.1.23
langchain-anthropic>=0.1.1
langchain-community>=0.0.13